* beta5:
  Rework honeypot code
  Update src/Services/CategoryService.php
  REFACTOR CategoryFacet - make category check more performant
  REFACTOR CategoryFacet - change code
  CHANGE CategoryFacet.php - facets of type 'category' are now available on category view
  FIX check for existing category before redirecting legacy routes
  add condition return empty array typ empty
  Add honeypot check to CustomerPasswordResetResource
  Add honeypot check and code formatting to CustomerResource
  Add honeypot check to contact-mail resource
  Add honeypot to destroy of CustomerNewsletterResource
  Add honeypot handling to CustomerNewsletterResource
  CHANGE ConfirmationController - cherrypick fix for hotfix release

# Conflicts:
#	src/Api/Resources/ContactMailResource.php
#	src/Api/Resources/CustomerResource.php
#	src/Extensions/Facets/CategoryFacet.php
#	src/Services/CategoryService.php

+++ src/Api/Resources/CustomerNewsletterResource.php
@@ -27,6 +27,13 @@ class CustomerNewsletterResource extends ApiResource
      */
     public function store(): Response
     {
+        // Honeypot check
+        if(strlen($this->request->get('honeypot')))
+        {
+            // We can potentially expand on the honeypot handling with sending reports to spam protect services
+            return $this->response->create('', ResponseCode::OK);
+        }
+
         $email = $this->request->get('email', '');
         $firstName = $this->request->get('firstName', '');
         $lastName = $this->request->get('lastName', '');
@@ -45,6 +52,12 @@ class CustomerNewsletterResource extends ApiResource
      */
     public function destroy(string $selector): Response
     {
+        // Honeypot check
+        if(strlen($this->request->get('honeypot')))
+        {
+            return $this->response->create(true, ResponseCode::OK);
+        }
+
         $emailFolder = $this->request->get('emailFolder', 0);
 
         $success = $this->newsletterService->deleteNewsletterDataByEmail($selector, (int) $emailFolder);
